Refactor: controller: create controld_cib.h
... with functions defined in controld_cib.c, for logical organization
